Part 1: ShipFileManagementSystem moved
