+++ Content.Client/_NF/Shipyard/BUI/ShipyardConsoleBoundUserInterface.cs
@@ -4,15 +4,8 @@ using Content.Shared._NF.Shipyard.BUI;
 using Content.Shared._NF.Shipyard.Events;
 using static Robust.Client.UserInterface.Controls.BaseButton;
 using Robust.Client.UserInterface;
-using Content.Client.Shuttles.Save;
+using Content.Client._Triad.Shipyard.Save;
 using Robust.Client.UserInterface.Controls;
-using Robust.Shared.IoC;
-using Robust.Shared.Log;
-using System.Linq;
-using Robust.Shared.Serialization.Markdown;
-using System.IO;
-using Robust.Shared.Serialization.Markdown.Mapping;
-using Robust.Shared.Serialization.Markdown.Value;
 using Robust.Shared.Configuration;
 using Content.Shared._Triad.CCVar;
